moved file location. example function to use the userApi
